# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CyberCat",
    "description": "Threat-informed automated incident response platform",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/events": {
      "get": {
        "tags": [
          "ingest"
        ],
        "summary": "List Events",
        "operationId": "list_events_v1_events_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/events/raw": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest Raw Event",
        "operationId": "ingest_raw_event_v1_events_raw_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RawEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RawEventAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/entities/{entity_id}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get Entity",
        "operationId": "get_entity_v1_entities__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entities": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Lookup Entity",
        "operationId": "lookup_entity_v1_entities_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityKind"
            }
          },
          {
            "name": "natural_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Natural Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntityDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/EntityKind"
          },
          "natural_key": {
            "type": "string",
            "title": "Natural Key"
          },
          "attrs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attrs"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "recent_events": {
            "items": {
              "$ref": "#/components/schemas/EntityTimelineEvent"
            },
            "type": "array",
            "title": "Recent Events"
          },
          "related_incidents": {
            "items": {
              "$ref": "#/components/schemas/EntityIncidentSummary"
            },
            "type": "array",
            "title": "Related Incidents"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "natural_key",
          "attrs",
          "first_seen",
          "last_seen",
          "recent_events",
          "related_incidents"
        ],
        "title": "EntityDetail"
      },
      "EntityIncidentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "$ref": "#/components/schemas/IncidentKind"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "confidence": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "kind",
          "status",
          "severity",
          "confidence",
          "opened_at",
          "updated_at"
        ],
        "title": "EntityIncidentSummary"
      },
      "EntityKind": {
        "type": "string",
        "enum": [
          "user",
          "host",
          "ip",
          "process",
          "file",
          "observable"
        ],
        "title": "EntityKind"
      },
...
      "EntityTimelineEvent": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "normalized": {
            "additionalProperties": true,
            "type": "object",
            "title": "Normalized"
          }
        },
        "type": "object",
        "required": [
          "id",
          "occurred_at",
          "kind",
          "normalized"
        ],
        "title": "EntityTimelineEvent"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
      "EventList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "EventList"
      },
      "EventSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "dedupe_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedupe Key"
          }
        },
        "type": "object",
        "required": [
          "id",
          "occurred_at",
          "source",
          "kind"
        ],
        "title": "EventSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentKind": {
        "type": "string",
        "enum": [
          "identity_compromise",
          "endpoint_compromise",
          "identity_endpoint_chain",
          "unknown"
        ],
        "title": "IncidentKind"
      },
...
      "IncidentStatus": {
        "type": "string",
        "enum": [
          "new",
          "triaged",
          "investigating",
          "contained",
          "resolved",
          "closed",
          "reopened"
        ],
        "title": "IncidentStatus"
      },
...
      "RawEventAccepted": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "dedup_hit": {
            "type": "boolean",
            "title": "Dedup Hit"
          },
          "detections_fired": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Detections Fired"
          },
          "incident_touched": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Touched"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "dedup_hit"
        ],
        "title": "RawEventAccepted"
      },
      "RawEventIn": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "direct",
              "seeder"
            ],
            "title": "Source"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "raw": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw"
          },
          "normalized": {
            "additionalProperties": true,
            "type": "object",
            "title": "Normalized"
          },
          "dedupe_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedupe Key"
          }
        },
        "type": "object",
        "required": [
          "source",
          "kind",
          "occurred_at",
          "raw",
          "normalized"
        ],
        "title": "RawEventIn"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "info",
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}